requirements.txt
dvc==3.50.1
fastapi==0.110.2
kaggle==1.6.14
matplotlib==3.8.4
mlflow==2.12.2
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0
prometheus-client==0.20.0
pydantic==2.6.4
python-multipart==0.0.9
scikit-learn==1.4.2
torch==2.2.2
torchvision==0.17.2
uvicorn==0.29.0
pytest==8.2.0
requests==2.31.0
